+++ lib/indexed/entity.ex
@@ -3,24 +3,24 @@ defmodule Indexed.Entity do
   defstruct fields: [], id_key: :id, lookups: [], prefilters: [], ref: nil
 
   @typedoc """
-  * `:fields` - List of `t:field/0`s for which pairs of lists should be
+  - `fields` : List of `t:field/0`s for which pairs of lists should be
     maintained with the ID sorted ascending and descending.
-  * `:id_key` - Specifies how to find the id for a record.  It can be an atom
+  - `id_key` : Specifies how to find the id for a record.  It can be an atom
     field name to access, a function, or a tuple in the form `{module,
     function_name}`. In the latter two cases, the record will be passed in.
     Default `:id`.
-  * `:lookups` - List of entity field names which should be indexed.
+  - `lookups` : List of entity field names which should be indexed.
     These fields can be leveraged via `Indexed.get_by/4`.
-  * `:prefilters` - List of tuples indicating which fields should be
+  - `prefilters` : List of tuples indicating which fields should be
     prefiltered on. This means that separate indexes will be managed for each
     unique value for each of these fields, across all records of this entity
     type. Each two-element tuple has the field name atom and a keyword list
     of options. Allowed options:
-    * `:maintain_unique` - List of field name atoms for which a list of
+    - `maintain_unique` : List of field name atoms for which a list of
       unique values under the prefilter will be managed. These lists can be
       fetched via `Indexed.get_uniques_list/4` and
       `Indexed.get_uniques_map/4`.
-  * `:ref` - ETS table reference where records of this entity type are
+  - `ref` : ETS table reference where records of this entity type are
     stored, keyed by id. This will be nil in the version compiled into a managed
     module.
   """
@@ -37,9 +37,9 @@ defmodule Indexed.Entity do
 
   ## Options
 
-  * `:sort` - Indicates how the field should be sorted in ascending order:
-    * `:datetime` - `DateTime.compare/2` should be used for sorting.
-    * `nil` (default) - `Enum.sort/1` will be used.
+  - `:sort` : Indicates how the field should be sorted in ascending order:
+    - `:datetime` : `DateTime.compare/2` should be used for sorting.
+    - `nil` (default) : `Enum.sort/1` will be used.
   """
   @type field :: {field_name, opts :: keyword}